クエリ関係のバグ修正
