+++ front/components/atom/Independent.tsx
@@ -1,26 +1,20 @@
 //lib
-import { IconPlusCircle } from '@supabase/ui'
 import Image, { StaticImageData } from 'next/image'
-
-//components
-import { PrimaryButton } from './PrimaryButton'
 
 interface Props {
   heading: string
   tips: string
   img: StaticImageData
-  isAdmin?: boolean
 }
 
-export const Independent: React.FC<Props> = ({ heading, tips, img, isAdmin }) => {
+export const Independent: React.FC<Props> = ({ heading, tips, img }) => {
   return (
     <div className="hero">
       <div className="hero-content text-center">
         <div className="mt-16 tracking-wider">
           <p className="text-4xl font-bold  text-gray-500">{heading}</p>
           <p className="py-6 text-xl text-gray-700">{tips}</p>
-          <Image src={img} width={400} height={390} />
-          {isAdmin && <PrimaryButton icon={<IconPlusCircle />} buttonText="グループを作成" />}
+          <Image src={img} width={400} height={390} alt="image" />
         </div>
       </div>
     </div>
