+++ front/pages/_app.tsx
@@ -1,10 +1,12 @@
 //lib
 import type { AppProps } from 'next/app'
+import { Suspense, useEffect, useState } from 'react'
 import { QueryClient, QueryClientProvider } from 'react-query'
 import { ReactQueryDevtools } from 'react-query/devtools'
 
 //components
 import { Notice } from '../components/atom/Notice'
+import { Spinner } from '../components/atom/Spinner'
 
 //style
 import '../styles/globals.css'
@@ -20,15 +22,17 @@ const queryClient = new QueryClient({
 })
 
 function MyApp({ Component, pageProps }: AppProps) {
-  return (
+  const [mounted, setMounted] = useState(false)
+  useEffect(() => setMounted(true), [])
+  return mounted ? (
     <QueryClientProvider client={queryClient}>
-      <>
+      <Suspense fallback={<Spinner />}>
         <Notice />
         <Component {...pageProps} />
-      </>
+      </Suspense>
       <ReactQueryDevtools initialIsOpen={false} />
     </QueryClientProvider>
-  )
+  ) : null
 }
 
 export default MyApp
