+++ front/pages/index.tsx
@@ -1,4 +1,5 @@
 //lib
+import { ExclamationCircleIcon } from '@heroicons/react/solid'
 import { NextPage } from 'next'
 
 //utils
@@ -7,31 +8,23 @@ import useStore from '../store'
 //components
 import { Auth } from '../components/Auth'
 import { DashBoard } from '../components/DashBoard'
-import { ErrorBoundary } from 'react-error-boundary'
-import { ExclamationCircleIcon } from '@heroicons/react/solid'
 import { Layout } from '../components/Layout'
-import { Suspense } from 'react'
-import { Spinner } from '../components/atom/Spinner'
+import { useEffect } from 'react'
+import { supabase } from '../utils/supabase'
 
 const Home: NextPage = () => {
   const session = useStore((state) => state.session)
+  const setSession = useStore((state) => state.setSession)
+
+  useEffect(() => {
+    setSession(supabase.auth.session())
+    supabase.auth.onAuthStateChange((_event, session) => {
+      setSession(session)
+    })
+  }, [setSession])
 
   const title = session ? 'ホーム' : 'ログイン'
-  return (
-    <Layout title={title}>
-      {!session ? (
-        <Auth />
-      ) : (
-        <ErrorBoundary
-          fallback={<ExclamationCircleIcon className="my-5 h-10 w-10 text-pink-500" />}
-        >
-          <Suspense fallback={<Spinner />}>
-            <DashBoard />
-          </Suspense>
-        </ErrorBoundary>
-      )}
-    </Layout>
-  )
+  return <Layout title={title}>{!session ? <Auth /> : <DashBoard />}</Layout>
 }
 
 export default Home
