+++ front/pages/setting/[uid].tsx
@@ -3,8 +3,15 @@ import { NextPage } from 'next'
 
 //components
 import { Layout } from '../../components/Layout'
+import { useQueryProfile } from '../../hooks/query/useQueryProfile'
 
 const Setting: NextPage = () => {
-  return <Layout title="設定">設定ページ</Layout>
+  const { data } = useQueryProfile()
+
+  return (
+    <Layout title="設定">
+      <p>{data?.username}さんの設定ページ</p>
+    </Layout>
+  )
 }
 export default Setting
